Refactor: move all markdown shenanigans to a separate file
